make hook a private action
