+++ packages/editor/src/private-apis.js
@@ -15,7 +15,6 @@ import PostPanelRow from './components/post-panel-row';
 import PostViewLink from './components/post-view-link';
 import PreviewDropdown from './components/preview-dropdown';
 import PreferencesModal from './components/preferences-modal';
-import useSaveEntities from './components/entities-saved-states/hooks/use-save-entities';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -31,7 +30,7 @@ lock( privateApis, {
 	PostViewLink,
 	PreviewDropdown,
 	PreferencesModal,
-	useSaveEntities,
+
 	// This is a temporary private API while we're updating the site editor to use EditorProvider.
 	useBlockEditorSettings,
 } );
